requirements.txt
keyboard~=0.13.5
colorama~=0.4.6
disnake~=2.9.2
requests~=2.32.4
python-dotenv~=1.0.1